+++ cosid-core/src/main/java/me/ahoo/cosid/util/ProcessId.java
@@ -14,6 +14,7 @@
 package me.ahoo.cosid.util;
 
 import javax.annotation.concurrent.Immutable;
+import java.lang.management.ManagementFactory;
 
 /**
  * get current process id .
@@ -21,17 +22,36 @@ import javax.annotation.concurrent.Immutable;
  * @author ahoo wang
  */
 @Immutable
-public final class ProcessId {
-
-    public static final ProcessId CURRENT = new ProcessId();
-
+public enum ProcessId {
+    CURRENT;
+    
     private final int processId;
-
-    public ProcessId() {
-        this.processId = Systems.getCurrentProcessId();
+    
+    ProcessId() {
+        this.processId = getCurrentProcessId();
     }
-
+    
     public int getProcessId() {
         return processId;
+    }
+    
+    /**
+     * get current process name .
+     *
+     * @return process name
+     */
+    public static String getCurrentProcessName() {
+        return ManagementFactory.getRuntimeMXBean().getName();
+    }
+    
+    /**
+     * get current process id .
+     *
+     * @return process id
+     */
+    public static int getCurrentProcessId() {
+        String processName = getCurrentProcessName();
+        String processIdStr = processName.split("@")[0];
+        return Integer.parseInt(processIdStr);
     }
 }
