+++ gradle.properties
@@ -12,7 +12,7 @@
 #
 
 group=me.ahoo.cosid
-version=1.8.10
+version=1.8.11
 
 description=Global distributed ID generator
 website=https://github.com/Ahoo-Wang/CosId
